Don't try to regenerate env.rb if it is read-only (as on Heroku)

+++ lib/bundler/environment.rb
@@ -76,12 +76,16 @@ module Bundler
     end
 
     def write_rb_lock
-      shared_helpers = File.read(File.expand_path("../shared_helpers.rb", __FILE__))
-      template = File.read(File.expand_path("../templates/environment.erb", __FILE__))
-      erb = ERB.new(template, nil, '-')
-      Bundler.env_file.dirname.mkpath
-      File.open(Bundler.env_file, 'w') do |f|
-        f.puts erb.result(binding)
+      if Bundler.env_file.exist? && !Bundler.env_file.writable?
+        Bundler.ui.warn "Cannot write to outdated .bundle/environment.rb to update it"
+      else
+        shared_helpers = File.read(File.expand_path("../shared_helpers.rb", __FILE__))
+        template = File.read(File.expand_path("../templates/environment.erb", __FILE__))
+        erb = ERB.new(template, nil, '-')
+        Bundler.env_file.dirname.mkpath
+        File.open(Bundler.env_file, 'w') do |f|
+          f.puts erb.result(binding)
+        end
       end
     end
